requirements.txt
streamlit
pandas
plotly
python-dateutil
pytest
nopenpyxl
kaleido
reportlab